chore: Move Babel extractor to new file, to not upset CI
