+++ pyproject.toml
@@ -28,4 +28,4 @@ name = "credere"
 version = "0.0.0"
 
 [project.entry-points."babel.extractors"]
-enum = "app.i18n:extract_enum"
+enum = "app.babel:extract_enum"
